requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
jedi==0.17.2
parso==0.7.1
pluggy==0.13.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
ujson==4.0.2
urllib3==1.26.4
